demos/demo5-secure-rag/pyproject.toml
[project]
name = "demo5-secure-rag"
version = "0.1.0"
description = "Demo 5: Secure RAG Agent — Hardening the Forecast Agent (Evolution of Demo 3)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flock-core",
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
leak-server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
